\name{compendiumdb-package}
\alias{compendiumdb-package}
\alias{compendiumdb}
\docType{package}
\title{
Tools for Retrieval and Storage of Functional Genomics Data}
\description{
  Public repositories such as the Gene Expression Omnibus (GEO) contain
  thousands of high-throughput functional genomics datasets. These
  datasets are a rich source of useful biological
  information. Extraction of meaningful information often requires the
  integration of a large number of datasets from different
  studies and platforms. The package compendiumdb provides a flexible
  platform for the systematic retrieval and storage of functional
  genomics data downloaded from GEO in the form of a MySQL database
  accessed via R functions. It provides functions to (i) download data
  from GEO, (ii) store data in the database and (iii) retrieve data from
  the database.
}
\details{
\tabular{ll}{
Package: \tab compendiumdb\cr
Type: \tab Package\cr
Version: \tab 1.0.3\cr
Date: \tab 2015-10-12\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Umesh K. Nandal <u.k.nandal@amc.uva.nl> and Perry D. Moerland
  <p.d.moerland@amc.uva.nl>
}
\keyword{ package }
