\name{compendiumdb-package}
\alias{compendiumdb-package}
\alias{compendiumdb}
\docType{package}
\title{
A database and R package for storing and analyzing gene expression data}
\description{
Public repositories such as Gene Expression Omnibus (GEO) contain thousands of microarray experiment 
datasets. These datasets are a rich source of useful biological information. Extraction of meaningful 
information often requires the integration of a large number of datasets from different microarray 
studies and platforms. The package compendiumdb provides a flexible platform for the systematic collection, 
storage and retrieval of gene expression data downloaded from GEO in the form of a MySQL database accessed 
via R functions. It provides functions to (a) download data from GEO and load data into the 
database, (b) store data in the database and (c) retrieve data from the database.
}
\details{
\tabular{ll}{
Package: \tab compendiumdb\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-02-13\cr
License: \tab GNU General Public License (GPL) - version 2\cr
LazyLoad: \tab yes\cr
}
}
\author{Umesh Nandal <u.k.nandal@amc.uva.nl>, Perry Moerland
  <p.d.moerland@amc.uva.nl>
}
\keyword{ package }
